+++ src/HttpApplicationRunner.php
@@ -24,11 +24,13 @@ use Yiisoft\Http\Header;
 use Yiisoft\Http\Method;
 use Yiisoft\Http\Status;
 use Yiisoft\PsrEmitter\EmitterInterface;
-use Yiisoft\PsrEmitter\HeadersHaveBeenSentException;
+use Yiisoft\PsrEmitter\HeadersHaveBeenSentException as EmitterHeadersHaveBeenSentException;
 use Yiisoft\PsrEmitter\SapiEmitter;
 use Yiisoft\Yii\Http\Application;
 use Yiisoft\Yii\Http\Handler\ThrowableHandler;
 use Yiisoft\Yii\Runner\ApplicationRunner;
+
+use Yiisoft\Yii\Runner\Http\Exception\HeadersHaveBeenSentException;
 
 use function in_array;
 use function microtime;
@@ -212,7 +214,11 @@ final class HttpApplicationRunner extends ApplicationRunner
         $response = $this->removeBodyByStatusMiddleware($response);
         $response = $this->contentLengthMiddleware($response);
         $response = $this->headRequestMiddleware($request, $response);
-        $this->emitter->emit($response);
+        try {
+            $this->emitter->emit($response);
+        } catch (EmitterHeadersHaveBeenSentException) {
+             throw new HeadersHaveBeenSentException();
+        }
     }
 
     /**